requirements.txt
tensorflow==2.11.0
keras==2.11.0
nltk==3.7
pillow==8.4.0
numpy==1.21.4
scikit-learn==0.24.2
matplotlib==3.4.3
tqdm==4.62.3
python-dotenv==0.19.2
rouge-score==0.0.4
pandas
ipykernel
ipywidgets
